backend/requirements.txt
fastapipip 

uvicorn[standard]
rapidfuzz
